+++ lib/src/folding/foldable_block.dart
@@ -55,6 +55,30 @@ class FoldableBlock extends InclusiveRange with EquatableMixin {
 
     return false;
   }
+
+  bool isSameLines(FoldableBlock other) {
+    return firstLine == other.firstLine && lastLine == other.lastLine;
+  }
+
+  FoldableBlock join(FoldableBlock other) {
+    return FoldableBlock(
+      firstLine: min(firstLine, other.firstLine),
+      lastLine: max(lastLine, other.lastLine),
+      type: _getJoinType(this, other),
+    );
+  }
+
+  static FoldableBlockType _getJoinType(FoldableBlock a, FoldableBlock b) {
+    if (a.type == FoldableBlockType.imports) {
+      return FoldableBlockType.imports;
+    }
+
+    if (b.type == FoldableBlockType.imports) {
+      return FoldableBlockType.imports;
+    }
+
+    return FoldableBlockType.union;
+  }
 }
 
 extension FoldableBlockList on List<FoldableBlock> {
@@ -62,31 +86,76 @@ extension FoldableBlockList on List<FoldableBlock> {
     sort((a, b) => a.firstLine - b.firstLine);
   }
 
-  /// Joins intersecting blocks in list.
-  /// Expected that list is sorted by start line.
+  /// Joins intersecting blocks.
+  ///
+  /// This list must be sorted by firstLine.
   void joinIntersecting() {
     if (length < 2) {
       return;
     }
 
-    for (int i = 1; i < length; i++) {
-      final currentBlock = this[i];
-      final previousBlock = this[i - 1];
+    // A working list to lay down blocks to compare.
+    // As we iterate blocks, it will contain the current nesting hierarchy.
+    final ancestors = <FoldableBlock>[];
+    final ancestorIndexToOverallIndex = <int>[];
 
-      final areIntersected = previousBlock.lastLine >= currentBlock.firstLine &&
-          previousBlock.lastLine < currentBlock.lastLine;
+    for (int overallIndex = 0; overallIndex < length; overallIndex++) {
+      // The pointer to the block we currently compare with ancestors.
+      // If it merges with any ancestor, it is redefined to point
+      // to that ancestor (bubbles up).
+      FoldableBlock bubble = this[overallIndex];
 
-      final isDuplicate = previousBlock.firstLine == currentBlock.firstLine &&
-          previousBlock.lastLine == currentBlock.lastLine;
+      // Throw the block into the current nesting hierarchy...
+      ancestors.add(bubble);
+      ancestorIndexToOverallIndex.add(overallIndex);
 
-      if (areIntersected || isDuplicate) {
-        this[i - 1] = FoldableBlock(
-          firstLine: previousBlock.firstLine,
-          lastLine: currentBlock.lastLine,
-          type: FoldableBlockType.union,
-        );
-        removeAt(i);
-        i = max(i - 2, 0);
+      int bubbleIndexInAncestors = ancestors.length - 1;
+      int bubbleOverallIndex = overallIndex;
+
+      // And fix every violation of the hierarchy by bubbling the block up,
+      // removing non-ancestors from the working list, and joining when needed.
+      for (int ancestorIndex = ancestors.length - 2;
+          ancestorIndex >= 0;
+          ancestorIndex--) {
+        final ancestor = ancestors[ancestorIndex];
+
+        if (ancestor.lastLine < bubble.firstLine) {
+          // `bubble` is not nested in `ancestor`.
+          // Remove `ancestor` from the working list and try the upper one.
+          ancestors.removeAt(ancestorIndex);
+          ancestorIndexToOverallIndex.removeAt(ancestorIndex);
+          bubbleIndexInAncestors--;
+          continue;
+        }
+
+        final isDuplicate = bubble.isSameLines(ancestor);
+
+        final areIntersecting = ancestor.lastLine >= bubble.firstLine &&
+            ancestor.lastLine < bubble.lastLine;
+
+        if (isDuplicate || areIntersecting) {
+          final joined = ancestor.join(bubble);
+
+          this[ancestorIndexToOverallIndex[ancestorIndex]] = joined;
+          ancestors[ancestorIndex] = joined;
+
+          removeAt(bubbleOverallIndex);
+          ancestors.removeAt(bubbleIndexInAncestors);
+          ancestorIndexToOverallIndex.removeAt(bubbleIndexInAncestors);
+          overallIndex--;
+
+          if (!areIntersecting) {
+            // `bubble` was a duplicate lines-wise.
+            // Do not go up the hierarchy because it is laid down alright there.
+            break; // to the top level, try the next block.
+          }
+
+          // Continue bubbling up the new joined block
+          // because it may still violate the hierarchy above.
+          bubbleIndexInAncestors = ancestorIndex;
+          bubbleOverallIndex = ancestorIndexToOverallIndex[ancestorIndex];
+          bubble = ancestors[bubbleIndexInAncestors];
+        }
       }
     }
   }