+++ lib/src/single_line_comments/parser/single_line_comments.dart
@@ -2,6 +2,7 @@ import 'package:highlight/highlight_core.dart';
 import 'package:highlight/languages/dart.dart';
 import 'package:highlight/languages/go.dart';
 import 'package:highlight/languages/java.dart';
+import 'package:highlight/languages/php.dart';
 import 'package:highlight/languages/python.dart';
 import 'package:highlight/languages/scala.dart';
 
@@ -9,11 +10,12 @@ class SingleLineComments {
   const SingleLineComments._();
 
   static final byMode = <Mode, List<String>>{
-    java: [_slashes],
     dart: [_slashes],
     go: [_slashes],
-    scala: [_slashes],
+    java: [_slashes],
+    php: [_slashes, _hash],
     python: [_hash],
+    scala: [_slashes],
   };
 
   static const _slashes = '//';