+++ test/src/common/text_editing_value.dart
@@ -5,12 +5,4 @@ extension TextEditingValueTestExtension on TextEditingValue {
   TextEditingValue replacedSelection(String text) {
     return replaced(selection, text);
   }
-
-  TextEditingValue typed(String text) {
-    final lengthDiff = text.length - selected.length;
-
-    return replaced(selection, text).copyWith(
-      selection: TextSelection.collapsed(offset: selection.end + lengthDiff),
-    );
-  }
 }